+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 * `all-is-cubes` library:
     * `block::BlockDef` no longer implements `AsRef<Block>`.
     * `block::Modifier::Attributes` has been replaced by `SetAttribute`.
+    * `space::PackedLight` no longer implements `From<Rgb>`. There is currently no replacement.
 
 
 ## 0.10.0 (2025-12-11)